upload image and generate invoice

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator/check');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', { 
@@ -12,9 +13,20 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.body.image;
     const price = req.body.price;
     const description = req.body.description;
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', { 
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            errorMessage: 'Attached file is not an image.',
+            product: { title: title, price: price, description: description },
+            validationErrors: []
+        });
+    }
 
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -23,10 +35,12 @@ exports.postAddProduct = (req, res, next) => {
             path: '/admin/add-product',
             editing: false,
             errorMessage: errors.array()[0].msg,
-            product: { title: title, imageUrl: imageUrl, price: price, description: description },
+            product: { title: title, price: price, description: description },
             validationErrors: errors.array()
         });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({title: title, price: price, description: description, imageUrl: imageUrl, userId: req.user});
     product.save()
@@ -68,18 +82,18 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.body.image;
     const price = req.body.price;
     const description = req.body.description;
 
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', { 
-            pageTitle: 'Add Product',
-            path: '/admin/add-product',
+            pageTitle: 'Edit Product',
+            path: '/admin/edit-product',
             editing: true,
             errorMessage: errors.array()[0].msg,
-            product: { _id: prodId, title: title, imageUrl: imageUrl, price: price, description: description },
+            product: { _id: prodId, title: title, price: price, description: description },
             validationErrors: errors.array()
         });
     }
@@ -89,7 +103,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = title;
             product.price = price;
             product.description = description;
-            product.imageUrl = imageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             return product.save()
                 .then(result => {
                     res.redirect('/admin/products');
@@ -122,7 +139,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.find({_id: prodId, userId: req.user._id})
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: prodId, userId: req.user._id});
+        })
         .then(result => {
             res.redirect('/admin/products');
         })
